Update T3Es_wrapper.py
Bug fix for the embedding - when the record header contains slashes it causes the output to be inside inner directories, and thus is discarded in downstream processes. I changed the record header to hold just the ID, and thus I guarantee (as much as possible at this point) that it contains no slashes.